Disable virtual keyboard and use breeze as the default theme.
The patch included - submitted upstream - allows the customization of default
settings (InputMethod and default theme) in configuration time - sddm
does not ship any config file by default, it embeds the defaults in code.
Then use that to disable the virtual keyboard and set Plasma's breeze theme
(as sddm is the dm in desktop-kde).

Related to distribution issues #380 and #381.